fix(cranelift-codegen): try to fix stable std CI

+++ cranelift/codegen/src/ir/immediates.rs
@@ -1971,7 +1971,7 @@ mod tests {
     fn fcvt_to_sint_negative_overflow_ieee64() {
         for n in &[8, 16, 32] {
             assert_eq!(-((1u64 << (n - 1)) as f64) - 1.0, unsafe {
-                mem::transmute::<_, f32>(Ieee64::fcvt_to_sint_negative_overflow(*n))
+                mem::transmute::<_, f64>(Ieee64::fcvt_to_sint_negative_overflow(*n))
             });
         }
     }
